build(deps): bump framels to version 0.7.7
Bump the version of framels from 0.7.6 to 0.7.7. This includes updates to the Cargo.lock and Cargo.toml files, as well as changes to the benches/mega.rs and docs/content/_index.md files. The updated version can be found on GitHub at [this link](https://github.com/forticheprod/fls/tree/0.7.7).